+++ Sources/Entities/ConnectionType.swift
@@ -1,0 +1,13 @@
+//
+//  ConnectionType.swift
+//  Tapglue
+//
+//  Created by John Nilsen on 8/2/16.
+//  Copyright © 2016 Tapglue. All rights reserved.
+//
+
+import Foundation
+
+public enum ConnectionType: String {
+    case Follow = "follow", Friend = "friend"
+}
